refactor: introduce factory and Swap class
