+++ implementations/cryptohauntological_probe/workers/bedrock_worker.py
