+++ implementations/cryptohauntological_probe/workers/claude_worker.py
