+++ implementations/cryptohauntological_probe/workers/cohere_worker.py
