+++ implementations/cryptohauntological_probe/workers/gemini_worker.py
