+++ implementations/cryptohauntological_probe/workers/mistral_worker.py
