+++ implementations/cryptohauntological_probe/workers/openai_worker.py
