+++ implementations/cryptohauntological_probe/workers/tinyllama_ollama_worker.py
